refactor: tls config validation now extracted to util function.
